[DX-2835] remove CRIB from the local CRE (#21114)
* remove CRIB

* go mod tidy
