+++ system-tests/lib/cre/features/consensus/v2/consensus.go
@@ -196,47 +196,6 @@ func formatBootstrapPeer(bootstrap *cre.Node) string {
 		cre.OCRPeeringPort)
 }
 
-func proposeBootstrapJob(creEnv *cre.Environment, bootstrap *cre.Node, don *cre.Don) (map[string][]string, error) {
-	input := cre_jobs.ProposeJobSpecInput{
-		Domain:      offchain.ProductLabel,
-		Environment: cre.EnvironmentName,
-		DONName:     bootstrap.DON.Name,
-		JobName:     "consensus-v2-bootstrap-" + don.Name,
-		ExtraLabels: map[string]string{cre.CapabilityLabelKey: flag},
-		DONFilters: []offchain.TargetDONFilter{
-			{Key: offchain.FilterKeyDONName, Value: bootstrap.DON.Name},
-		},
-		Template: job_types.BootstrapOCR3,
-		Inputs: job_types.JobSpecInput{
-			"chainSelector":     creEnv.RegistryChainSelector,
-			"contractQualifier": ContractQualifier,
-		},
-	}
-
-	proposer := cre_jobs.ProposeJobSpec{}
-	if verErr := proposer.VerifyPreconditions(*creEnv.CldfEnvironment, input); verErr != nil {
-		return nil, fmt.Errorf("precondition verification failed for Consensus v2 bootstrap job: %w", verErr)
-	}
-
-	report, applyErr := proposer.Apply(*creEnv.CldfEnvironment, input)
-	if applyErr != nil {
-		return nil, fmt.Errorf("failed to propose Consensus v2 bootstrap job spec: %w", applyErr)
-	}
-
-	specs := make(map[string][]string)
-	for _, r := range report.Reports {
-		out, ok := r.Output.(cre_jobs_ops.ProposeOCR3BootstrapJobOutput)
-		if !ok {
-			return nil, fmt.Errorf("unable to cast to ProposeOCR3BootstrapJobOutput, actual type: %T", r.Output)
-		}
-		if mergeErr := mergo.Merge(&specs, out.Specs, mergo.WithAppendSlice); mergeErr != nil {
-			return nil, fmt.Errorf("failed to merge bootstrap job specs: %w", mergeErr)
-		}
-	}
-
-	return specs, nil
-}
-
 func proposeNodeJob(creEnv *cre.Environment, don *cre.Don, command string, bootstrapPeers []string, configStr string) (map[string][]string, error) {
 	capRegVersion, ok := creEnv.ContractVersions[keystone_changeset.CapabilitiesRegistry.String()]
 	if !ok {
