+++ system-tests/tests/.gitignore
@@ -15,8 +15,6 @@ cre_v*
 
 .cre.settings.yaml
 
-# crib-specific CL node configuration and secrets that will be used to override the default configuration
-crib-configs/
 cache/
 blockscout/
 compose/
