--- system-tests/tests/load/cre/writer-don-load-test-crib.toml
@@ -1,74 +1,0 @@
-# ----- BLOCKCHAIN CONFIGURATION -----
-# Sets up a local Anvil blockchain for testing with 5-second block time
-[[blockchains]]
-chain_id = "1337"
-type = "anvil"
-docker_cmd_params = ["-b", "5"]
-
-# ----- JOB DISTRIBUTOR -----
-# Docker image for the job distributor service - hardcoded version
-[jd]
-csa_encryption_key = "d1093c0060d50a3c89c189b2e485da5a3ce57f3dcb38ab7e2c0d5f0bb2314a44" # any random 32 byte hex string
-image = "job-distributor:0.22.1"
-
-# ----- MOCK CAPABILITIES -----
-# Required mock trigger capability for connecting to the write target
-[[mock_capabilities]]
-name="mock"
-version= "1.0.0"
-type="trigger"
-description="mock"
-
-# ----- TEST CONFIGURATION -----
-# Parameters for the Writer DON load test
-[writer_test]
-nr_of_feeds = 30                # Number of price feeds to simulate
-workflow_name = "abcdef0123"    # Must be exactly 10 characters
-workflow_owner = "0x0100000000000000000000000000000000000001"
-workflow_id = "5dbe5f217ff07d6b1dddb43519fe7bf13ccb10b540578fafdbea86b508abbd71"
-
-# ----- INFRASTRUCTURE SETUP -----
-# Configures AWS-based CRIB infrastructure
-[infra]
-type = "crib"
-
-# CRIB deployment configuration for AWS
-[infra.crib]
-namespace = "crib-writer-don-load-test"
-folder_location = "../../../../../crib/deployments/cre"  # Path to CRIB deployment folder
-provider = "aws"
-
-# Team information for resource allocation and billing
-[infra.crib.team_input]
-team="cre"
-product="writer-don"
-component="writer-don"
-cost_center="capabilities-df-migration-performance-tests"
-
-# ----- NODE CONFIGURATION -----
-# Setup for a 6-node Writer DON cluster (includes bootstrap node)
-[[nodesets]]
-nodes = 6
-override_mode = "all"           # Apply same configuration to all nodes
-http_port_range_start = 10100   # Starting port for HTTP endpoints
-name = "writer"
-
-# Database configuration for the nodes
-[nodesets.db]
-image = "postgres:12.0"
-port = 13000
-volume_name = "writer_volume"
-
-# ----- NODE SPECIFICATIONS -----
-# Chainlink node configuration with required features enabled
-[[nodesets.node_specs]]
-[nodesets.node_specs.node]
-image = "public.ecr.aws/chainlink/chainlink:2.23.0-ubuntu20" # Chainlink node image for AWS deployment
-user_config_overrides = """
-        [Feature]
-        LogPoller = true
-        FeedsManager = true
-        [OCR2]
-        Enabled = true
-        DatabaseTimeout = '1s'
-      """
